+++ src/human-readable/index.ts
@@ -1,3 +1,20 @@
+export {
+  CircularReferenceError,
+  InvalidParenthesisError,
+  UnknownSignatureError,
+  InvalidSignatureError,
+  InvalidStructSignatureError,
+  InvalidAbiParameterError,
+  InvalidAbiParametersError,
+  InvalidParameterError,
+  SolidityProtectedKeywordError,
+  InvalidModifierError,
+  InvalidFunctionModifierError,
+  InvalidAbiTypeParameterError,
+  InvalidAbiItemError,
+  UnknownTypeError,
+} from './errors'
+
 export { parseAbi, type ParseAbi } from './parseAbi'
 
 export { parseAbiItem, type ParseAbiItem } from './parseAbiItem'